src/domain/services/ai/ai_integration.py
"""AI integration service implementation.

This module implements the AI service interface.
It provides functionality for:
- Route optimization
- Price predictions
- Natural language processing
- Model integration
"""
from datetime import datetime
from typing import Dict, List, Optional, Tuple, Union
from uuid import UUID

from src.domain.entities.route import Route
from src.domain.interfaces.services.ai_service import AIService
from src.domain.services.common.base import BaseService
from src.domain.value_objects import (
    AIModelResponse,
    OptimizationResult,
    PricePrediction,
    RouteOptimization
)

class AIIntegrationService(BaseService, AIService):
    """Service for AI model integration.
    
    This service is responsible for:
    - Route optimization using AI models
    - Price prediction using ML models
    - Natural language query processing
    - Model selection and versioning
    """
    
    def __init__(
        self,
        model_client: Optional['AIModelClient'] = None,
        cache_service: Optional['CacheService'] = None
    ):
        """Initialize AI integration service.
        
        Args:
            model_client: Optional client for AI model API
            cache_service: Optional service for caching
        """
        super().__init__()
        self._model_client = model_client
        self._cache_service = cache_service
        self._retry_count = 3
        self._cache_ttl = 3600  # 1 hour
    
    def optimize_route(
        self,
        route: Route,
        optimization_type: str = "cost",
        constraints: Optional[Dict] = None
    ) -> RouteOptimization:
        """Optimize route using AI models.
        
        Args:
            route: Route to optimize
            optimization_type: Type of optimization
            constraints: Optional constraints
            
        Returns:
            Optimization results
            
        Raises:
            ValueError: If optimization fails
        """
        self._log_entry(
            "optimize_route",
            route=route,
            optimization_type=optimization_type
        )
        
        try:
            # Check cache first
            cache_key = f"route_opt_{route.id}_{optimization_type}"
            if self._cache_service:
                cached = self._cache_service.get(cache_key)
                if cached:
                    return RouteOptimization(**cached)
            
            # Prepare optimization request
            request = {
                "route_data": route.dict(),
                "optimization_type": optimization_type,
                "constraints": constraints or {}
            }
            
            # Call AI model
            response = self._call_model_with_retry(
                "route_optimization",
                request
            )
            
            if not response:
                raise ValueError("Failed to get optimization response")
            
            # Process results
            optimization = RouteOptimization(
                route_id=route.id,
                original_distance=route.distance_km,
                optimized_distance=response.get("distance"),
                original_duration=route.duration_hours,
                optimized_duration=response.get("duration"),
                optimization_type=optimization_type,
                improvements=response.get("improvements", []),
                score=response.get("score"),
                metadata={
                    "model_version": response.get("model_version"),
                    "optimization_time": response.get("processing_time"),
                    "confidence": response.get("confidence")
                }
            )
            
            # Cache results
            if self._cache_service:
                self._cache_service.set(
                    cache_key,
                    optimization.dict(),
                    self._cache_ttl
                )
            
            self._log_exit("optimize_route", optimization)
            return optimization
            
        except Exception as e:
            self._log_error("optimize_route", e)
            raise ValueError(f"Route optimization failed: {str(e)}")
    
    def predict_price(
        self,
        route: Route,
        cargo_type: Optional[str] = None,
        market_factors: Optional[Dict] = None
    ) -> PricePrediction:
        """Predict optimal price using ML models.
        
        Args:
            route: Route for prediction
            cargo_type: Optional cargo type
            market_factors: Optional market factors
            
        Returns:
            Price prediction
            
        Raises:
            ValueError: If prediction fails
        """
        self._log_entry(
            "predict_price",
            route=route,
            cargo_type=cargo_type
        )
        
        try:
            # Check cache first
            cache_key = f"price_pred_{route.id}_{cargo_type}"
            if self._cache_service:
                cached = self._cache_service.get(cache_key)
                if cached:
                    return PricePrediction(**cached)
            
            # Prepare prediction request
            request = {
                "route_data": route.dict(),
                "cargo_type": cargo_type,
                "market_factors": market_factors or {}
            }
            
            # Call AI model
            response = self._call_model_with_retry(
                "price_prediction",
                request
            )
            
            if not response:
                raise ValueError("Failed to get prediction response")
            
            # Process results
            prediction = PricePrediction(
                route_id=route.id,
                predicted_price=response.get("price"),
                price_range=(
                    response.get("min_price"),
                    response.get("max_price")
                ),
                confidence=response.get("confidence"),
                factors=response.get("factors", []),
                metadata={
                    "model_version": response.get("model_version"),
                    "prediction_time": response.get("processing_time"),
                    "market_conditions": response.get("market_conditions")
                }
            )
            
            # Cache results
            if self._cache_service:
                self._cache_service.set(
                    cache_key,
                    prediction.dict(),
                    self._cache_ttl
                )
            
            self._log_exit("predict_price", prediction)
            return prediction
            
        except Exception as e:
            self._log_error("predict_price", e)
            raise ValueError(f"Price prediction failed: {str(e)}")
    
    def process_query(
        self,
        query: str,
        context: Optional[Dict] = None
    ) -> AIModelResponse:
        """Process natural language query.
        
        Args:
            query: Query to process
            context: Optional context
            
        Returns:
            Model response
            
        Raises:
            ValueError: If processing fails
        """
        self._log_entry("process_query", query=query)
        
        try:
            # Check cache first
            cache_key = f"query_{hash(query)}_{hash(str(context))}"
            if self._cache_service:
                cached = self._cache_service.get(cache_key)
                if cached:
                    return AIModelResponse(**cached)
            
            # Prepare query request
            request = {
                "query": query,
                "context": context or {}
            }
            
            # Call AI model
            response = self._call_model_with_retry(
                "query_processing",
                request
            )
            
            if not response:
                raise ValueError("Failed to get query response")
            
            # Process results
            result = AIModelResponse(
                query=query,
                response=response.get("answer"),
                confidence=response.get("confidence"),
                suggestions=response.get("suggestions", []),
                metadata={
                    "model_version": response.get("model_version"),
                    "processing_time": response.get("processing_time"),
                    "context_used": response.get("context_used")
                }
            )
            
            # Cache results
            if self._cache_service:
                self._cache_service.set(
                    cache_key,
                    result.dict(),
                    self._cache_ttl
                )
            
            self._log_exit("process_query", result)
            return result
            
        except Exception as e:
            self._log_error("process_query", e)
            raise ValueError(f"Query processing failed: {str(e)}")
    
    def _call_model_with_retry(
        self,
        endpoint: str,
        request: Dict
    ) -> Optional[Dict]:
        """Call AI model with retry logic.
        
        Args:
            endpoint: Model endpoint
            request: Request data
            
        Returns:
            Model response or None
            
        Raises:
            ValueError: If all retries fail
        """
        if not self._model_client:
            raise ValueError("No AI model client configured")
        
        last_error = None
        for attempt in range(self._retry_count):
            try:
                return self._model_client.call(endpoint, request)
            except Exception as e:
                last_error = e
                self.logger.warning(
                    f"Model call failed (attempt {attempt + 1}): {str(e)}",
                    exc_info=True
                )
                if attempt < self._retry_count - 1:
                    import time
                    time.sleep(2 ** attempt)  # Exponential backoff
        
        if last_error:
            raise ValueError(f"All retries failed: {str(last_error)}")
        return None
